update loading page
