+++ index.html
@@ -18,8 +18,6 @@
     <meta name="theme-color" content="#000000">
 </head>
 <body>
-    <div id="dev-loading" style="font-family: monospace; font-size: 14px; color: white; background: black; height: 100vh">
-        Loading...
-    </div>
+    <div id="root"></div>
 </body>
 </html>
