+++ package.json
@@ -69,6 +69,8 @@
         "prettier": "^2.7.1",
         "react-custom-scrollbars-2": "^4.4.0",
         "react-helmet": "^6.1.0",
+        "react-helmet-async": "^1.3.0",
+        "react-loader-spinner": "^5.4.5",
         "react-router-dom": "^6.7.0",
         "terser": "^5.15.1",
         "unplugin-swc": "^1.3.2",
