+++ src/components/head-component/HeadComponent.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Helmet } from 'react-helmet';
+import { Helmet, HelmetProvider } from 'react-helmet-async';
 
 interface IProps {
-    children: JSX.Element;
+    children: React.ReactElement;
     title: string;
     description: string;
     image: string;
@@ -11,14 +11,16 @@ interface IProps {
 export default function HeadComponent(props: IProps) {
     return(
         <>
-            <Helmet>
-                <title>{props.title}</title>
-                <meta name="description" content={props.description}/>
-                <meta property="og:title" content={props.title}/>
-                <meta property="og:description" content={props.description}/>
-                <meta property="og:image" content={props.image}/>
-            </Helmet>
-            {props.children}
+            <HelmetProvider>
+                <Helmet>
+                    <title>{props.title}</title>
+                    <meta name="description" content={props.description}/>
+                    <meta property="og:title" content={props.title}/>
+                    <meta property="og:description" content={props.description}/>
+                    <meta property="og:image" content={props.image}/>
+                </Helmet>
+                {props.children}
+            </HelmetProvider>
         </>
     );
 }
